refactor: simplify usePanZoom hook by extracting by concern
